including the complete and failed download broadcasts into the thread
